begin split of G2phsGUI
